Schema updates for easier scaling
Updated schemas to be easier scalable. A contributor now extends a user as the data is exactly the same. The contributor just also has a tally of how many contribution they have made. A repository now has an owner object. This owner object just follows the user schema. As an owner of a repo also is just a user.